added background and starting appBar
