+++ public/manifest.json
@@ -1,5 +1,5 @@
 {
-    "short_name": "YM",
+    "short_name": "YM's Site",
     "name": "Yehuda Makarov",
     "icons": [
         {
