+++ src/Root/Root.tsx
@@ -1,28 +1,11 @@
-import React from "react";
-import { Backdrop, Typography, makeStyles, createStyles, Paper, Theme } from "@material-ui/core";
+import React from 'react'
+import { Button } from "@material-ui/core"
+import TopBar from "./TopBar/TopBar";
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles(
-        {
-            root: {
-                backgroundColor: theme.palette.background.paper,
-                bottom: 0,
-                right: 0,
-                top: 0,
-                left: 0,
-                position: "fixed"
-            }
-        }
-    )
-)
-
-function Root() {
-    const classes = useStyles();
+export default function Root() {
     return (
         <React.Fragment>
-            <Paper className={classes.root} />
+            <TopBar />
         </React.Fragment>
     )
 }
-
-export default Root
